GFETCH-87 Item click on SupportedORder recycler view has been implemented. Order code of the item is also fetched, now need to implement how to determine the workflow according the order code
