+++ app/src/main/java/co/gobd/gofetch/adapter/SupportedOrderAdapter.java
@@ -6,6 +6,8 @@ import android.view.InflateException;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.TextView;
 
 import com.squareup.picasso.Picasso;
 
@@ -14,19 +16,19 @@ import java.util.List;
 
 import co.gobd.gofetch.R;
 import co.gobd.gofetch.callback.OnOrderDataReceive;
+import co.gobd.gofetch.listener.OnItemClickListener;
 import co.gobd.gofetch.model.SupportedOrder;
 import co.gobd.gofetch.service.SupportedOrderService;
-import co.gobd.gofetch.view.SupportedOrderViewHolder;
 
 
 /**
  * Created by tonmoy on 20-Jan-16.
  */
-public class SupportedOrderAdapter extends RecyclerView.Adapter<SupportedOrderViewHolder> implements OnOrderDataReceive {
+public class SupportedOrderAdapter extends RecyclerView.Adapter<SupportedOrderAdapter.SupportedOrderViewHolder> implements OnOrderDataReceive {
 
     private List<SupportedOrder> itemList;
     private Context context;
-
+    private OnItemClickListener onItemClickListener;
 
     public SupportedOrderAdapter(Context context) {
         this.context = context;
@@ -37,11 +39,17 @@ public class SupportedOrderAdapter extends RecyclerView.Adapter<SupportedOrderVi
         service.loadSupportedOrder();
     }
 
+    // Any UI component that needs to know the List item click and position needs to set this listener
+    public void setOnItemClickListener(final OnItemClickListener onItemClickListener) {
+        this.onItemClickListener = onItemClickListener;
+    }
+
     @Override
     public SupportedOrderViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         SupportedOrderViewHolder viewHolder = null;
         try {
-            View layoutView = LayoutInflater.from(parent.getContext()).inflate(R.layout.cardview_supported_order, null);
+            View layoutView = LayoutInflater.from(parent.getContext())
+                    .inflate(R.layout.cardview_supported_order, null);
             viewHolder = new SupportedOrderViewHolder(layoutView);
         } catch (InflateException e) {
             e.printStackTrace();
@@ -57,14 +65,11 @@ public class SupportedOrderAdapter extends RecyclerView.Adapter<SupportedOrderVi
         if (!itemList.isEmpty()) {
             holder.tvActionName.setText(itemList.get(position).getActionName());
             holder.tvItemName.setText(itemList.get(position).getOrderName());
-
             // Downloads and updates the image view with Picasso
             Picasso
                     .with(context)
                     .load(itemList.get(position).getImageUrl())
                     .into(holder.ivItemImage);
-
-
         }
     }
 
@@ -77,9 +82,40 @@ public class SupportedOrderAdapter extends RecyclerView.Adapter<SupportedOrderVi
 
     @Override
     public void onReceive(List<SupportedOrder> supportedOrders) {
-
         // Carries the data from OrderService and updates the data set
         this.itemList = supportedOrders;
         notifyDataSetChanged();
     }
+
+    public class SupportedOrderViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
+
+        public ImageView ivItemImage;
+        public TextView tvActionName;
+        public TextView tvItemName;
+
+        public SupportedOrderViewHolder(View itemView) {
+            super(itemView);
+            itemView.setOnClickListener(this);
+
+            ivItemImage = (ImageView) itemView.findViewById(R.id.iv_order);
+            tvActionName = (TextView) itemView.findViewById(R.id.tv_action);
+            tvItemName = (TextView) itemView.findViewById(R.id.tv_item);
+
+        }
+
+        @Override
+        public void onClick(View view) {
+            // Gets the clicked item's position in the view
+            int position = getLayoutPosition();
+            if (!itemList.isEmpty()) {
+                String orderCode = itemList.get(position).getOrderCode();
+                // TODO Find a way to decide which activity/workflow needs to be initiated according to the orderCode
+                if (onItemClickListener != null) {
+                    onItemClickListener.onClick(view, position);
+                }
+            }
+        }
+    }
+
+
 }
