+++ app/src/main/java/co/gobd/gofetch/listener/OnItemClickListener.java
@@ -1,0 +1,11 @@
+package co.gobd.gofetch.listener;
+
+import android.view.View;
+
+/**
+ * Created by tonmoy on 23-Mar-16.
+ * Handles click events on RecyclerView
+ */
+public interface OnItemClickListener {
+    void onClick(View view, int position);
+}
